Add support for external MongoDB instances & new Rongo.auth() function
Added support for use of external MongoDB instances using the Rongo Web API, which will be released shortly after this commit. 

Additionally, added a new Rongo.auth() function which allows for simpler setup for auth, directly in the Rongo.new() method, with support for email/password, API key, Bearer and now connection string/URI.

The only caveat to external instances is the API needed on top, however this is a simple fix by deploying the Rongo Web API to your server.